# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Hedera KMS Signing API
  description: >
    Serverless Hedera transaction signing backend on AWS.
    Submit CryptoTransfer and HTS token transfer signing requests without
    handling private keys. AWS KMS holds the ECDSA secp256k1 signing key,
    Lambda orchestrates the signing flow, and DynamoDB provides an immutable
    audit trail.
  version: "1.0.0"
  contact:
    name: Hedera KMS Signer
  license:
    name: MIT
servers:
  - url: https://{apiId}.execute-api.{region}.amazonaws.com
    description: AWS API Gateway endpoint
    variables:
      apiId:
        default: "xxxxxxxxxx"
        description: API Gateway ID
      region:
        default: us-east-1
        description: AWS region
security:
  - bearerAuth: []
paths:
  /sign-transfer:
    post:
      operationId: signTransfer
      summary: Submit a CryptoTransfer signing request
      description: >
        Validates the request, evaluates policy rules, builds a Hedera
        CryptoTransfer transaction, signs it via AWS KMS, and submits it
        to the Hedera network. Supports idempotency via client-generated
        requestId — duplicate requests with the same payload return the
        cached result; duplicate requestId with a different payload returns
        409 Conflict.
      tags:
        - Signing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignTransferRequest"
            example:
              requestId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              senderAccountId: "0.0.8291501"
              recipientAccountId: "0.0.1234"
              amountHbar: 1.5
              memo: "Payment for services"
      responses:
        "200":
          description: Transaction signed and submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignTransferResponse"
              example:
                transactionId: "0.0.8291501@1719000000.000000000"
                status: "SUCCESS"
                transactionHash: "a3f1b2c4d5e6f7890123456789abcdef0123456789abcdef0123456789abcdef"
        "400":
          description: Validation error — invalid or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              example:
                error: "VALIDATION_ERROR"
                details:
                  - "requestId must be a valid UUID"
                  - "amountHbar must be a positive number"
        "401":
          description: Missing or invalid authentication credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
              example:
                message: "Unauthorized"
        "403":
          description: Policy denied — request violates one or more policy rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDeniedError"
              example:
                error: "POLICY_DENIED"
                violations:
                  - "AMOUNT_EXCEEDS_MAX"
                  - "RECIPIENT_NOT_ALLOWED"
        "409":
          description: Conflict — requestId reused with a different payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
              example:
                error: "CONFLICT"
                message: "requestId already used with a different payload"
        "502":
          description: Signing or submission error from upstream service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
              examples:
                signingError:
                  summary: KMS signing failure
                  value:
                    error: "SIGNING_ERROR"
                    message: "KMS Sign API call failed after 3 attempts"
                submissionError:
                  summary: Hedera submission failure
                  value:
                    error: "SUBMISSION_ERROR"
                    message: "Transaction submission timed out after 30s"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
              example:
                error: "INTERNAL_ERROR"
  /public-key:
    get:
      operationId: getPublicKey
      summary: Retrieve KMS public key info
      description: >
        Returns the KMS-managed public key in multiple formats: DER-encoded,
        compressed (33 bytes), uncompressed (65 bytes), and the derived EVM
        address. The public key is cached after the first KMS API call.
      tags:
        - Key Info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Public key information retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyResponse"
              example:
                publicKeyDer: "3056301006072a8648ce3d020106052b8104000a03420004bfcab..."
                publicKeyCompressed: "02bfcab37d42f5a3d7162a0fd38e8d0fedf0a5e6b3c8d9e1f2a3b4c5d6e7f8091"
                publicKeyUncompressed: "04bfcab37d42f5a3d7162a0fd38e8d0fedf0a5e6b3c8d9e1f2a3b4c5d6e7f80911a2b3c4d5e6f7081929a3b4c5d6e7f80911a2b3c4d5e6f7081929a3b4c5d6e7f8"
                evmAddress: "0x1a2b3c4d5e6f70819293a4b5c6d7e8f901234567"
        "401":
          description: Missing or invalid authentication credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
              example:
                message: "Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
              example:
                error: "INTERNAL_ERROR"
  /rotate-key:
    post:
      operationId: rotateKey
      summary: Initiate signing key rotation
      description: >
        Creates a new KMS signing key, updates the Hedera account key list
        to include the new public key, and disables the old key. Uses the
        KMS_KEY_ID and HEDERA_OPERATOR_ID environment variables to identify
        the current key and operator account. No request body is needed.
      tags:
        - Key Management
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Key rotation completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateKeyResponse"
              example:
                oldKeyId: "alias/hedera-signer-dev"
                newKeyId: "arn:aws:kms:us-east-1:123456789:key/new-key-uuid"
                hederaTransactionId: "0.0.8291501@1234567890.000"
                status: "ROTATION_COMPLETE"
        "401":
          description: Missing or invalid authentication credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
              example:
                message: "Unauthorized"
        "500":
          description: Key rotation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotationError"
              example:
                error: "ROTATION_ERROR"
                message: "KMS CreateKey did not return a KeyId"
  /docs:
    get:
      operationId: getDocs
      summary: Serve OpenAPI specification
      description: >
        Returns this OpenAPI 3.0 specification as YAML. Useful for importing
        into Postman, Swagger UI, or other API documentation tools.
      tags:
        - Documentation
      security: []
      responses:
        "200":
          description: OpenAPI specification returned successfully
          content:
            text/yaml:
              schema:
                type: string
              example: "openapi: '3.0.3'\ninfo:\n  title: Hedera KMS Signing API\n  ..."
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
              example:
                error: "INTERNAL_ERROR"
  /sign-token-transfer:
    post:
      operationId: signTokenTransfer
      summary: Submit an HTS token transfer signing request
      description: >
        Validates the request, evaluates policy rules, builds a Hedera Token
        Service (HTS) transfer transaction, signs it via AWS KMS, and submits
        it to the Hedera network. Uses the same keccak256 + KMS signing bridge
        as CryptoTransfer. Supports idempotency via client-generated requestId.
      tags:
        - Signing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignTokenTransferRequest"
            example:
              requestId: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
              senderAccountId: "0.0.8291501"
              recipientAccountId: "0.0.1234"
              tokenId: "0.0.456789"
              amount: 100
              memo: "Token transfer via KMS"
      responses:
        "200":
          description: Token transfer signed and submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignTransferResponse"
              example:
                transactionId: "0.0.8291501@1719000000.000000000"
                status: "SUCCESS"
                transactionHash: "b4f2c3d5e6f7890123456789abcdef0123456789abcdef0123456789abcdef01"
        "400":
          description: Validation error — invalid or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              example:
                error: "VALIDATION_ERROR"
                details:
                  - "tokenId must be a valid Hedera token ID (0.0.N)"
                  - "amount must be a positive integer (smallest token unit)"
        "401":
          description: Missing or invalid authentication credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Policy denied — request violates one or more policy rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDeniedError"
        "409":
          description: Conflict — requestId reused with a different payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "502":
          description: Signing or submission error from upstream service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
  /multisig-config:
    get:
      operationId: getMultiSigConfig
      summary: Retrieve multi-signature configuration
      description: >
        Returns the current multi-signature / threshold key configuration.
        Shows whether multi-sig is enabled, the threshold, total keys, and
        details about each key (label, signing method, public key).
      tags:
        - Key Management
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Multi-sig configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSigConfigResponse"
              example:
                enabled: false
                threshold: 1
                totalKeys: 1
                keys:
                  - label: "KMS Primary"
                    publicKeyHex: "(derived at runtime from KMS)"
                    signingMethod: "kms"
                    kmsKeyId: "alias/hedera-signer-dev"
                description: "Single-key mode. The KMS key is the sole signer."
        "401":
          description: Missing or invalid authentication credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Bearer token authentication. For MVP, a JWT issued by Amazon Cognito
        or a shared bearer token configured in API Gateway.
  schemas:
    SignTransferRequest:
      type: object
      required:
        - requestId
        - senderAccountId
        - recipientAccountId
        - amountHbar
      properties:
        requestId:
          type: string
          format: uuid
          description: Client-generated UUID for idempotency
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        senderAccountId:
          type: string
          pattern: "^0\\.0\\.[1-9]\\d*$"
          description: "Hedera sender account ID (format: 0.0.N)"
          example: "0.0.8291501"
        recipientAccountId:
          type: string
          pattern: "^0\\.0\\.[1-9]\\d*$"
          description: "Hedera recipient account ID (format: 0.0.N)"
          example: "0.0.1234"
        amountHbar:
          type: number
          minimum: 0
          exclusiveMinimum: true
          maximum: 5
          description: Amount in HBAR to transfer (positive, max 5 for MVP)
          example: 1.5
        memo:
          type: string
          maxLength: 100
          description: Optional transaction memo
          example: "Payment for services"
...
    UnauthorizedError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Unauthorized error message
...